Simplified StateSubject solution
